eachContainsEach(), toArray(), toArray(T[] a)
eachContains each is a new utility. It takes a Set and an Array, and checks that all the elements in each are in the other.

new toArray edge test. Checks that everything works as planned when mySet hold objects of multiply types.

Normal test for toArray(T[] a). Passes to large an array, and then asserts that each el has been added to the array in the correct order, and that excess space is filled with nulls. Asserts also that all els in the original set are in the return.

edge test for the same checks if the passed array was too small to hold all els, and returns a properly sized one is so. Then runs a eachContainsEach() test on the set and the array return.